+++ src/LaTeXBlocks.RenderHost/Properties/AssemblyInfo.cs
@@ -7,5 +7,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("Copyright 2026")]
 [assembly: ComVisible(false)]
 [assembly: Guid("A3ED3B72-57CD-4AC0-9DAC-D3DB21A2024E")]
-[assembly: AssemblyVersion("0.2.71.0")]
-[assembly: AssemblyFileVersion("0.2.71.0")]
+[assembly: AssemblyVersion("0.2.72.0")]
+[assembly: AssemblyFileVersion("0.2.72.0")]